PeerGroup: Move maxOfMostFreq() from Utils to here, as here is the only usage.
