+++ core/src/main/java/org/bitcoinj/core/PeerGroup.java
@@ -2624,7 +2624,40 @@ public class PeerGroup implements TransactionBroadcaster, GovernanceVoteBroadcas
             return 0;
         List<Integer> heights = new ArrayList<>(peers.size());
         for (Peer peer : peers) heights.add((int) peer.getBestHeight());
-        return Utils.maxOfMostFreq(heights);
+        return maxOfMostFreq(heights);
+    }
+
+    private static class Pair implements Comparable<Pair> {
+        int item, count;
+        public Pair(int item, int count) { this.count = count; this.item = item; }
+        // note that in this implementation compareTo() is not consistent with equals()
+        @Override public int compareTo(Pair o) { return -Integer.compare(count, o.count); }
+    }
+
+    static int maxOfMostFreq(List<Integer> items) {
+        if (items.isEmpty())
+            return 0;
+        // This would be much easier in a functional language (or in Java 8).
+        items = Ordering.natural().reverse().sortedCopy(items);
+        LinkedList<Pair> pairs = new LinkedList<>();
+        pairs.add(new Pair(items.get(0), 0));
+        for (int item : items) {
+            Pair pair = pairs.getLast();
+            if (pair.item != item)
+                pairs.add((pair = new Pair(item, 0)));
+            pair.count++;
+        }
+        // pairs now contains a uniqified list of the sorted inputs, with counts for how often that item appeared.
+        // Now sort by how frequently they occur, and pick the max of the most frequent.
+        Collections.sort(pairs);
+        int maxCount = pairs.getFirst().count;
+        int maxItem = pairs.getFirst().item;
+        for (Pair pair : pairs) {
+            if (pair.count != maxCount)
+                break;
+            maxItem = Math.max(maxItem, pair.item);
+        }
+        return maxItem;
     }
 
     /**
