+++ core/src/test/java/org/bitcoinj/core/UtilsTest.java
@@ -21,7 +21,6 @@ package org.bitcoinj.core;
 import java.math.BigInteger;
 import java.util.Arrays;
 import java.util.Date;
-import java.util.List;
 
 import org.junit.Test;
 
@@ -32,15 +31,6 @@ public class UtilsTest {
     @Test
     public void testReverseBytes() {
         assertArrayEquals(new byte[]{1, 2, 3, 4, 5}, Utils.reverseBytes(new byte[]{5, 4, 3, 2, 1}));
-    }
-
-    @Test
-    public void testMaxOfMostFreq() throws Exception {
-        assertEquals(0, Utils.maxOfMostFreq());
-        assertEquals(0, Utils.maxOfMostFreq(0, 0, 1));
-        assertEquals(2, Utils.maxOfMostFreq(1, 1, 2, 2));
-        assertEquals(1, Utils.maxOfMostFreq(1, 1, 2, 2, 1));
-        assertEquals(-1, Utils.maxOfMostFreq(-1, -1, 2, 2, -1));
     }
 
     @Test
